+++ magerun/execute.go
@@ -8,6 +8,7 @@ import (
 	"magecomm/logger"
 	"magecomm/notifictions"
 	"os/exec"
+	"regexp"
 	"strings"
 )
 
@@ -17,16 +18,30 @@ func HandleMagerunCommand(messageBody string) (string, error) {
 	command, args := parseMagerunCommand(messageBody)
 	args = sanitizeCommandArgs(args)
 
-	if !config_manager.IsMageRunCommandAllowed(command) {
-		return "", fmt.Errorf("command %s is not allowed", command)
+	if isCmdAllowed, err := config_manager.IsMageRunCommandAllowed(command); !isCmdAllowed {
+		return "", err
 	}
 
-	if config_manager.IsRestrictedCommandArgsIncluded(command, args) {
-		return "", fmt.Errorf("the command '%s' is not allowed with the following arguments: %s", command, strings.Join(args, " "))
+	if isRestrictedArgsIncluded, err := config_manager.IsRestrictedCommandArgsIncluded(command, args); isRestrictedArgsIncluded {
+		return "", err
 	}
 
 	if isAllRequiredArgsIncluded, missingRequiredArgs := config_manager.IsRequiredCommandArgsIncluded(command, args); !isAllRequiredArgsIncluded {
 		return "", fmt.Errorf("the command '%s' is missing some required arguments: %s, unable to run command", command, strings.Join(missingRequiredArgs, " "))
+	}
+
+	//if --no-interaction/-n is not set, set it
+	forceNoInteractionFlag := config_manager.GetBoolValue(config_manager.CommandConfigForceMagerunNoInteraction)
+	noInteractionFlagPresent := false
+	for _, arg := range args {
+		if arg == "--no-interaction" || arg == "-n" {
+			noInteractionFlagPresent = true
+			break
+		}
+	}
+	if !noInteractionFlagPresent && forceNoInteractionFlag {
+		logger.Infof("The command '%s' does not contain the '--no-interaction' flag, adding it to the command", command)
+		args = append(args, "--no-interaction")
 	}
 
 	args = append([]string{command}, args...)
@@ -35,7 +50,7 @@ func HandleMagerunCommand(messageBody string) (string, error) {
 
 func executeMagerunCommand(args []string) (string, error) {
 	mageRunCmdPath := getMageRunCommand()
-	logger.Infof("Executing command %s with args: %v\n", mageRunCmdPath, args)
+	logger.Infof("Executing command %s with args: %v", mageRunCmdPath, args)
 
 	if config_manager.GetBoolValue(config_manager.ConfigSlackEnabled) {
 		logger.Infof("Slack notification is enabled, sending notification")
@@ -43,7 +58,7 @@ func executeMagerunCommand(args []string) (string, error) {
 		err := notifier.Notify(
 			fmt.Sprintf("Executing command: '%v' on environment: '%s'", strings.Join(args, " "), config_manager.GetValue(config_manager.CommandConfigEnvironment)))
 		if err != nil {
-			logger.Warnf("Failed to send slack notification: %v\n", err)
+			logger.Warnf("Failed to send slack notification: %v", err)
 		}
 	}
 
@@ -56,16 +71,34 @@ func executeMagerunCommand(args []string) (string, error) {
 	cmd.Stderr = &stderrBuffer
 
 	err := cmd.Run()
-	if err != nil {
-		return "", fmt.Errorf("error executing magerun command: %s", err)
-	}
-
+	// Grab any output before returning with command error
 	stdoutStr := stdoutBuffer.String()
 	stderrStr := stderrBuffer.String()
+	output := stripMagerunOutput(stdoutStr + "\n" + stderrStr)
 
-	output := stdoutStr + "\n" + stderrStr
-
+	// Now check command for error and return either success or failure
+	if err != nil {
+		logger.Warnf("Error executing magerun command: %s, with the following output: %s", err, strings.ReplaceAll(output, "\n", " "))
+		return output, fmt.Errorf("error executing magerun command: %s", err)
+	}
 	return output, nil
+}
+
+func stripMagerunOutput(output string) string {
+	patterns := map[string]string{
+		`(?i)(?:it's|it is) not recommended to run .*? as root user`: "",
+		//Add more regex patterns here with their corresponding replacement
+	}
+
+	strippedOutput := output
+	for pattern, replacement := range patterns {
+		re := regexp.MustCompile(pattern)
+		strippedOutput = re.ReplaceAllString(strippedOutput, replacement)
+	}
+	//trim any leading or trailing whitespace
+	strippedOutput = strings.TrimSpace(strippedOutput)
+
+	return strippedOutput
 }
 
 func getMageRunCommand() string {
@@ -82,7 +115,7 @@ func parseMagerunCommand(messageBody string) (string, []string) {
 	return args[0], args[1:]
 }
 
-// We absolutely must not allow command escaping. e.g magerun cache:clean; rm -rf /
+// We absolutely must not allow command escaping. e.g. magerun cache:clean; rm -rf /
 func sanitizeCommandArgs(args []string) []string {
 	var sanitizedArgs []string
 	disallowed := []string{";", "&&", "||", "|", "`", "$", "(", ")", "<", ">", "!"}